feature/WEL-24_user-profile: moving Header and Footer from pages to components
