+++ src/pages/AboutUsPage/AboutUsPage.tsx
@@ -1,6 +1,5 @@
-
-import Footer from "../Footer/Footer";
-import Header from "../Header/Header";
+import Footer from "../../components/Footer/Footer";
+import Header from "../../components/Header/Header";
 
 const AboutUsPage: React.FC = () => {
   return (
